// common/enums/executor-type.ts
export enum ExecutorTypes {
  ANY = 'any',
  INDIVIDUAL = 'individual',
  COMPANY = 'company',
}

export enum ExecutorMenuTabs {
  MAIN = 'main',
  SERVICES = 'services',
  REQUISITES = 'requisites',
}

export enum ExecutorQuestionnaireTabs {
  MAIN = 'main',
  ABOUT_SERVICE = 'about_service',
  RATING = 'rating',
}